Prepend wildcards for nexus no-proxy entries
We take the unix-style no-proxy list and convert it to be suitable for
nexus. This gives us the ability to only have one source for a no-proxy
list, at the cost of not having the full power of the Java wildcard
system. For example, you can't have a wilcard in the middle like
`10.*.0`, because we're limited to the lowest common denominator of all
systems.