Fix failure to update the script
Concept
The shell script is executed by an interpreter and there is no assurance that the whole file is loaded to the memory from the disk during the execution. The interpreter will use the line numbers to load the next chunk of the file from disk to memory. If the file gets updated in the meantime, then the new chunk that has been loaded to the memory could lead to errors.

Solution
The interpreter will always respect the scope. Hence, bring the statements that could change the file within a single scope and always exit the execution before the scope ends.